Move Map files to src Folder
Moves the Map.cpp and Map.hpp files to the src folder.
